agent-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
langchain-core==0.3.39
langchain==0.3.6
langchain-community==0.3.4
langchain-text-splitters==0.3.1
langgraph==0.2.40
langsmith==0.1.147
langchain-google-genai==2.0.10
langserve==0.3.3
sse_starlette>=3.0.0
pydantic==2.7.4
pydantic-settings==2.11.0
email-validator==2.1.0
python-dotenv==1.0.1
httpx==0.27.0
loguru==0.7.2
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.5
sqlalchemy==2.0.44
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.9
redis==5.0.1
cryptography==42.0.0

pytest==8.3.3
